pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "product-extractor"
version = "0.1.0"
description = "Extract product blocks and structured fields from catalog PDFs"
readme = "README.md"
authors = [{name="4sito"}]
license = {text = "MIT"}
requires-python = ">=3.9"

dependencies = [
  "pymupdf>=1.22.0",
  "pillow>=9.0.0",
  "pytesseract>=0.3.10",
  "opencv-python-headless>=4.7.0",
  "numpy>=1.21",
  "pandas>=1.5"
]

[project.scripts]
product-extractor = "product_extractor.cli:main"
